+++ src/video_core/renderer_vulkan/vk_device.cpp
@@ -14,7 +14,7 @@
 #include "common/assert.h"
 #include "core/settings.h"
 #include "video_core/renderer_vulkan/vk_device.h"
-#include "video_core/renderer_vulkan/wrapper.h"
+#include "video_core/vulkan_common/vulkan_wrapper.h"
 
 namespace Vulkan {
