+++ src/video_core/renderer_vulkan/vk_master_semaphore.cpp
@@ -8,7 +8,7 @@
 #include "core/settings.h"
 #include "video_core/renderer_vulkan/vk_device.h"
 #include "video_core/renderer_vulkan/vk_master_semaphore.h"
-#include "video_core/renderer_vulkan/wrapper.h"
+#include "video_core/vulkan_common/vulkan_wrapper.h"
 
 namespace Vulkan {
