+++ src/video_core/renderer_vulkan/vk_master_semaphore.h
@@ -8,7 +8,7 @@
 #include <thread>
 
 #include "common/common_types.h"
-#include "video_core/renderer_vulkan/wrapper.h"
+#include "video_core/vulkan_common/vulkan_wrapper.h"
 
 namespace Vulkan {
