+++ src/video_core/renderer_vulkan/vk_shader_util.cpp
@@ -9,7 +9,7 @@
 #include "common/common_types.h"
 #include "video_core/renderer_vulkan/vk_device.h"
 #include "video_core/renderer_vulkan/vk_shader_util.h"
-#include "video_core/renderer_vulkan/wrapper.h"
+#include "video_core/vulkan_common/vulkan_wrapper.h"
 
 namespace Vulkan {
