+++ src/video_core/vulkan_common/vulkan_wrapper.cpp
@@ -13,7 +13,7 @@
 #include "common/common_types.h"
 #include "common/logging/log.h"
 
-#include "video_core/renderer_vulkan/wrapper.h"
+#include "video_core/vulkan_common/vulkan_wrapper.h"
 
 namespace Vulkan::vk {
